1. Separated Router functions 2. Implemented EJS view engine templates 3. Created shared views 4. Adopted MVC folder structure 5. Added Content Resources Folder 6. Added 'Colors' dependency ================================== Known Issues: 1. File directory displaying issue. jQuery is not successfully loaded in the home page. Seems need to enable static content displaying.
