+++ route/index.js
@@ -1,0 +1,12 @@
+module.exports.routers = [
+{
+    method: 'GET',
+    path: '/',
+    handler: function (request, reply) {
+        reply.view('./index/home', {
+        	title: 'Home Page',
+        	message: 'Index - Home Page'
+        });
+    }
+}
+];
