+++ route/static_content.js
@@ -1,0 +1,9 @@
+module.exports.routers = [
+	{
+		method: 'GET',
+		path: '/content',
+		handler: function(request, reply) {
+			reply.file('');
+		}
+	}
+]
