+++ view/_shared/footer.ejs
@@ -1,0 +1,2 @@
+</body>
+</html>
