+++ view/_shared/header.ejs
@@ -1,0 +1,5 @@
+<html lang="en">
+<head>
+	<title><%=title %></title>
+</head>
+<body>
