+++ view/_shared/javascripts.ejs
@@ -1,0 +1,1 @@
+<script src='/content/js/jquery-1.12.3.min.js'></script>
