+++ view/index/home.ejs
@@ -1,0 +1,12 @@
+<!-- Header Section -->
+<%- include('../_shared/header') %>
+
+<!-- Main Body Content -->
+<h1><%=message %> </h1>
+
+
+<!-- Javascript Implementation -->
+<%- include('../_shared/javascripts') %>
+
+<!-- Footer Section -->
+<%- include('../_shared/footer') %>
